shuffle the function_signature tests
This is necessary to reduce the maximum RAM spike when executing
make test-unit. If the function_signature tests are done in
alphabetical order, then several tests with similar names and
high RAM requirements will be executed at once if done in
parallel, which may cause the machine to swap or if swap is
exhausted / unavailable, to cause the tests to fail. By shuffling
these tests, it is much less likely to be executing many tests
with high RAM requirements at the same time.