grad_reg_inc_gamma: more accurate for large z
Compute the gradient of the asymptotic expansion of the
incomplete gamma, for use when z is large.

This was motivated by seeing inaccurate gradients in
gamma_ccdf_log(16,3,3), failing the testcase I was trying to add. This
was due to accumulated error in the series expansion when z is large.

Though the asymptotic expansion doesn't appear common (e.g. scipy uses
continued fractions instead), it is simple to differentiate.
Numerical accuracy was checked using sympy as a reference,
comparing the accuracy of the series expansion to the asymptotic
expansion. Over a grid (50 evenly-logspaced values of A between
0.001 and 1000), the cutoff (z>=a)&&(z>=8) was found to be a
sufficient condition.

Analysis:
http://nbviewer.jupyter.org/gist/b11z/4f49105acbe90cab29a9c3ae56590e4a